Made various improvements

+++ data/journals.js
@@ -1,6 +1,7 @@
 import { journals, users } from '../config/mongoCollections.js';
 import { ObjectId } from 'mongodb';
 import * as helpers from '../helpers.js';
+import * as sectionMethods from './sections.js';
 
 export const createJournal = async (userId, username, title) => {
   try {
@@ -67,6 +68,24 @@ export const updateJournal = async (journalId, updatedJournal) => {
 export const deleteJournal = async (journalId) => {
   journalId = helpers.checkString(journalId, "Journal ID");
   const journalCollection = await journals();
+  const journal = await journalCollection.findOne({_id: new ObjectId(journalId)});
+  if (!journal) throw "We could not find the journal to be deleted.";
+  const userCollection = await users();
+  const user = await userCollection.findOne({username: journal.author[0]});
+  if (!user) throw "We could not find the user that owns the journal.";
+  let section;
+  for (let i in journal.sections) {
+    section = await sectionMethods.deleteSection(journal.sections[i]);
+    if (!section) throw "We could not delete the section.";
+  }
+  for (let i in user.journals) {
+    if (user.journals[i] === journalId) {
+      user.journals[i] = user.journals[user.journals.length - 1];
+      user.journals.pop();
+    }
+  }
+  const updatedUser = await userCollection.findOneAndReplace({username: user.username}, user);
+  if (!updatedUser) throw "We could not update the user.";
   const deleteInfo = await journalCollection.deleteOne({ _id: new ObjectId(journalId) });
   if (deleteInfo.deletedCount === 0) {
     throw `Could not delete journal with id ${journalId}`;
